requirements.txt
requests~=2.31.0
pandas~=2.3.3
xlrd==2.0.1
openpyxl==3.1.5
pyarrow~=22.0.0
sqlalchemy~=2.0.20
psycopg2-binary~=2.9.9
python-dotenv~=1.1.1

# WRDS if we want variable descriptions for the tables...


## DBT and SQLite
#dbt-core==1.9.1
#dbt-sqlite==1.9.1